+++ tests/integration/init_test.go
@@ -19,7 +19,6 @@ package integration
 import (
 	"context"
 	"fmt"
-	"io/ioutil"
 	"os"
 	"path/filepath"
 	"testing"
@@ -34,7 +33,7 @@ import (
 func TestInit(t *testing.T) {
 	t.Parallel()
 
-	dir, err := ioutil.TempDir("", "")
+	dir, err := os.MkdirTemp("", "")
 	if err != nil {
 		t.Fatalf("unexpected err: %v", err)
 	}
@@ -96,7 +95,7 @@ func TestInitNewNoMerge(t *testing.T) {
 func testInitNew(t *testing.T, merge bool) {
 	clusterName := uuid.Must(uuid.NewV4()).String()
 
-	dir, err := ioutil.TempDir("", "")
+	dir, err := os.MkdirTemp("", "")
 	if err != nil {
 		t.Fatalf("unexpected err: %v", err)
 	}
@@ -175,7 +174,7 @@ func TestInitExistingNoMerge(t *testing.T) {
 func testInitExisting(t *testing.T, merge bool) {
 	clusterName := uuid.Must(uuid.NewV4()).String()
 
-	dir, err := ioutil.TempDir("", "")
+	dir, err := os.MkdirTemp("", "")
 	if err != nil {
 		t.Fatalf("unexpected err: %v", err)
 	}
@@ -307,7 +306,7 @@ func testInitExisting(t *testing.T, merge bool) {
 func TestInitUsers(t *testing.T) {
 	t.Parallel()
 
-	dir, err := ioutil.TempDir("", "")
+	dir, err := os.MkdirTemp("", "")
 	if err != nil {
 		t.Fatalf("unexpected err: %v", err)
 	}
@@ -413,7 +412,7 @@ func TestInitUsers(t *testing.T) {
 func TestInitialClusterSpec(t *testing.T) {
 	t.Parallel()
 
-	dir, err := ioutil.TempDir("", "")
+	dir, err := os.MkdirTemp("", "")
 	if err != nil {
 		t.Fatalf("unexpected err: %v", err)
 	}
@@ -467,7 +466,7 @@ func TestInitialClusterSpec(t *testing.T) {
 func TestExclusiveLock(t *testing.T) {
 	t.Parallel()
 
-	dir, err := ioutil.TempDir("", "")
+	dir, err := os.MkdirTemp("", "")
 	if err != nil {
 		t.Fatalf("unexpected err: %v", err)
 	}
@@ -523,7 +522,7 @@ func TestExclusiveLock(t *testing.T) {
 func TestPasswordTrailingNewLine(t *testing.T) {
 	t.Parallel()
 
-	dir, err := ioutil.TempDir("", "")
+	dir, err := os.MkdirTemp("", "")
 	if err != nil {
 		t.Fatalf("unexpected err: %v", err)
 	}