+++ tests/integration/standby_test.go
@@ -18,7 +18,6 @@ package integration
 
 import (
 	"fmt"
-	"io/ioutil"
 	"os"
 	"path/filepath"
 	"testing"
@@ -33,7 +32,7 @@ import (
 func TestInitStandbyCluster(t *testing.T) {
 	t.Parallel()
 
-	dir, err := ioutil.TempDir("", "stolon")
+	dir, err := os.MkdirTemp("", "stolon")
 	if err != nil {
 		t.Fatalf("unexpected err: %v", err)
 	}
@@ -98,7 +97,7 @@ func TestInitStandbyCluster(t *testing.T) {
 	storePath := filepath.Join(common.StorePrefix, clusterName)
 	sm := store.NewKVBackedStore(tstore.store, storePath)
 
-	pgpass, err := ioutil.TempFile(dir, "pgpass")
+	pgpass, err := os.CreateTemp(dir, "pgpass")
 	if err != nil {
 		t.Fatalf("unexpected err: %v", err)
 	}
@@ -166,7 +165,7 @@ func TestInitStandbyCluster(t *testing.T) {
 func TestPromoteStandbyCluster(t *testing.T) {
 	t.Parallel()
 
-	dir, err := ioutil.TempDir("", "stolon")
+	dir, err := os.MkdirTemp("", "stolon")
 	if err != nil {
 		t.Fatalf("unexpected err: %v", err)
 	}
@@ -231,7 +230,7 @@ func TestPromoteStandbyCluster(t *testing.T) {
 	storePath := filepath.Join(common.StorePrefix, clusterName)
 	sm := store.NewKVBackedStore(tstore.store, storePath)
 
-	pgpass, err := ioutil.TempFile(dir, "pgpass")
+	pgpass, err := os.CreateTemp(dir, "pgpass")
 	if err != nil {
 		t.Fatalf("unexpected err: %v", err)
 	}
@@ -310,13 +309,13 @@ func TestPromoteStandbyCluster(t *testing.T) {
 func TestPromoteStandbyClusterArchiveRecovery(t *testing.T) {
 	t.Parallel()
 
-	dir, err := ioutil.TempDir("", "stolon")
+	dir, err := os.MkdirTemp("", "stolon")
 	if err != nil {
 		t.Fatalf("unexpected err: %v", err)
 	}
 	defer os.RemoveAll(dir)
 
-	archiveBackupDir, err := ioutil.TempDir(dir, "archivebackup")
+	archiveBackupDir, err := os.MkdirTemp(dir, "archivebackup")
 	if err != nil {
 		t.Fatalf("unexpected err: %v", err)
 	}
@@ -384,7 +383,7 @@ func TestPromoteStandbyClusterArchiveRecovery(t *testing.T) {
 	storePath := filepath.Join(common.StorePrefix, clusterName)
 	sm := store.NewKVBackedStore(tstore.store, storePath)
 
-	pgpass, err := ioutil.TempFile(dir, "pgpass")
+	pgpass, err := os.CreateTemp(dir, "pgpass")
 	if err != nil {
 		t.Fatalf("unexpected err: %v", err)
 	}